requirements.txt
numpy==1.21.5
scipy==1.7.2
opencv-python==4.6.0.66
pillow==9.0.1
termcolor==1.1.0
pyyaml==6.0
tqdm==4.63.0
absl-py==1.2.0
gdown==4.5.1
ftfy
regex
git+https://github.com/openai/CLIP.git
transformers==4.25.1
PyMCubes==0.1.2
pymesh==1.0.2
pymeshlab
trimesh==3.17.1
scikit-image==0.19.3
scikit-learn==1.0.2
chumpy==0.70
neural_renderer_pytorch